move reusable functions from PropertiesPane.js into AbstractPropertiesList.js
